==> twitter_sentimental_analysis/Src/2_TweetAnalysis+Pre-Processing.ipynb <==
# Part 2 - Preprocessing Tweets

## 1. Importing Libraries
### 1.1 Libraries

pip install findspark

#Initializing pandas & numpy
import pandas as pd
import numpy as np

# Text Pre-processing libraries
import nltk
import re
from nltk.tokenize import word_tokenize
from string import punctuation 
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer,PorterStemmer
# Downloading from the collection
nltk.download('stopwords')
nltk.download('punkt')
nltk.download('wordnet')


#Spark Libraries
import findspark
from pyspark import SparkContext
from pyspark.sql import SparkSession, Window, Row
from pyspark.sql.functions import *
from pyspark.sql.types import *

# Emoji library
import emoji

#Data visualization 
from matplotlib import pyplot as plt

# Spark Session
spark = SparkSession.builder.appName("Spark").getOrCreate()

### 1.2 Configurations

# Directory to store the tweets data after analysis & pre-processing
tweet_df_path = "../Data/tweets_data.csv"
processed_tweet_df_path = "../Data/processed_tweet_data.csv" 

## 2. Loading Tweets

# Loading the data
tweet_df = spark.read.format('csv').options(header='true').option("mode", "DROPMALFORMED").load(tweet_df_path)

## 3. Analysing Data

# previewing the df table
tweet_df.limit(5).toPandas()

#Verifying the schema of the data table
tweet_df.printSchema()

#Converting spark df to pandas
tweet_df = tweet_df.toPandas()

#NaN values removal
tweet_df = tweet_df[tweet_df["tweet"].notna()]

# Adding index number 
tweet_df.index.names = ['Index_number']

#Viewing the data
tweet_df

## 4. Cleaning Tweets

#Initializing Lemm & Stem
lemmatizer = WordNetLemmatizer()
stemmer = PorterStemmer() 

#setting the english stopwords into the var stop_words + punctual
stop_words = set(stopwords.words('english')+ list(punctuation))

# Function to pre-process tweets
def process_tweets(text):
    """
    This function cleans the tweets by stripping off URLs, miscellaneous characters. 
    Then the tweets are tokenized, stemmed, and lemmatized. 
    
    Parameters:
    text: The tweet.
    Retrun: The processed tweet.
    """
   # Text cleaning
    text = text.lower()
    text = re.sub('((www\.[^\s]+)|(https?://[^\s]+))', ' ', text)
    text = re.sub(r'[^\x00-\x7F]+',' ', text)
    text = re.sub('@[^\s]+', ' ', text)
    text = emoji.demojize(text)
    text = word_tokenize(text)
    text = [word for word in text if word not in stop_words]
    text = [stemmer.stem(w) for w in text]
    text = [lemmatizer.lemmatize(w) for w in text]

    #joining every word in the text by a space
    return " ".join(text)


# Inside the tweet column, cleaned tweet(text)applied
tweet_df.tweet = tweet_df.tweet.apply(process_tweets)

#creating csv in the "processed_tweet_df_path" path 
tweet_df.to_csv(processed_tweet_df_path, index=False)
