==> twitter_sentimental_analysis/Src/3_Sentimental_Analysis.ipynb <==
# Part 3 - Sentiment Analysis 

# 1. Importing Libraries
### 1.1 Libraries

#Library to Access Amazon
import boto3

# Library to provide data structures(dataframes)
import pandas as pd
import numpy as np

### 1.2 Configuration

# Importing Credentials from config file
from configs import configs

# Authentication Amazon Comprehend
comprehend=boto3.client(
                         service_name="comprehend",  # Name of the Service
                         region_name="us-east-2", # region for the Service
                         aws_access_key_id=configs["amazon_credentials"]["AWS_SERVER_PUBLIC_KEY"], # Access Credentials
                         aws_secret_access_key=configs["amazon_credentials"]["AWS_SERVER_SECRET_KEY"] # Access Credentials
                       )

# Directory from pre-processing
processed_tweet_df_path = "../Data/processed_tweet_data.csv" 

# Directory to store the new data with sentiments
processed_tweet_df_path_with_sentiment = "../Data/processed_tweet_df_path_with_sentiment.csv"

### 1.3 Loading Dataframe

# Loading the data to dataframe
tweet_df = pd.read_csv(processed_tweet_df_path, index_col=False)

#Nan
tweet_df = tweet_df[tweet_df["tweet"].notna()]

## 2. Performing Sentiment Analysis
### 2.1 Defining Functions

def apply_comprehend(tweet):
    
    """
    This function uses amazon comprehend to predict the sentiment behind the tweet.
    
    Parameter:
    tweet: Each tweet.
    
    Return "result": A dictionary that contains the overall sentiment of the tweet, followed by the sentiment score under 
    Positive, Negative,Neutral, and mixed categories.    
    """
    
    try:
            out=comprehend.detect_sentiment(
                                             Text=tweet,
                                             LanguageCode="en"
                                           )
            #Dictionary
            result = {
                        "Sentiment": out["Sentiment"],
                        "SentimentScore_Positive": out["SentimentScore"]["Positive"],
                        "SentimentScore_Negative": out["SentimentScore"]["Negative"],
                        "SentimentScore_Neutral": out["SentimentScore"]["Neutral"],
                        "SentimentScore_Mixed": out["SentimentScore"]["Mixed"]  
                     }
            
            
            
    except Exception as e:
        # Incase of an error arises, nan values are returned. (Dictionary)
        result = {
                    "Sentiment": np.nan,
                    "SentimentScore_Positive": np.nan,
                    "SentimentScore_Negative": np.nan,
                    "SentimentScore_Neutral": np.nan,
                    "SentimentScore_Mixed": np.nan 
                 }

        print(f"Exception: {e}")
            
    return result

### 2.2 Performing Sentiment Analysis

#Time
from time import time
start_time = time()

#Using dataframe.apply function to apply the apply_comprehend function on each and every row of the dataframe.
#Storing the results into the dataframe under new columns.

tweet_df_with_sentiment = tweet_df.merge(
                                        tweet_df.tweet.apply(lambda x: (
                                                                        sentiment_score := apply_comprehend(x),


                                                                        pd.Series(
                                                                                      {
                                                                                       'Sentiment': sentiment_score["Sentiment"],
                                                                                       'SentimentScore_Positive': sentiment_score["SentimentScore_Positive"],
                                                                                       'SentimentScore_Negative': sentiment_score["SentimentScore_Negative"],
                                                                                       'SentimentScore_Neutral':sentiment_score["SentimentScore_Neutral"],
                                                                                       'SentimentScore_Mixed': sentiment_score["SentimentScore_Mixed"],
                                                                                      }
                                                                                  )
                                                                        )[-1]
                                                            ), 

                                         left_index=True,
                                         right_index=True
                                    )

#Time taken
print(f"\nTotal time = {time() - start_time} s")

# Adding index number 
tweet_df_with_sentiment.index.names = ['Index_number']

# Verifying the structure
tweet_df_with_sentiment.shape

### 2.3 Viewing Results

#Viewing the results
tweet_df_with_sentiment.sample(5)

## 3. Saving Dataframe

#creating csv in the "processed_tweet_df_path_with_sentiment" path 
tweet_df_with_sentiment.to_csv(processed_tweet_df_path_with_sentiment, index=None)
